+++ idp_engine/Annotate.py
@@ -27,13 +27,14 @@ from .Parse import (Vocabulary, Extern, TypeDeclaration,
                     Theory, Definition, Rule,
                     Structure, SymbolInterpretation, Enumeration, FunctionEnum,
                     Tuple, ConstructedFrom, Display)
-from .Expression import (Expression, Constructor, IfExpr, AQuantification,
+from .Expression import (Expression, Constructor, IfExpr, AQuantification, Quantee,
                          ARImplication, AImplication, AEquivalence, ADisjunction,
-                         Operator, AComparison, AUnary, AAggregate,
+                         AConjunction, Operator, AComparison, AUnary, AAggregate,
                          AppliedSymbol, UnappliedSymbol, Variable, Brackets,
                          FALSE, SymbolExpr, Number)
 
-from .utils import BOOL, INT, REAL, DATE, SYMBOL, OrderedSet, IDPZ3Error
+from .utils import (BOOL, INT, REAL, DATE, SYMBOL, OrderedSet, IDPZ3Error,
+                    DEF_SEMANTICS, Semantics)
 
 
 # Class Vocabulary  #######################################################
@@ -133,7 +134,7 @@ def annotate(self, theory, voc, q_vars):
     self.set_level_symbols()
 
     # create common variables, and rename vars in rule
-    self.clarks = {}
+    self.canonicals = {}
     for r in self.rules:
         decl = voc.symbol_decls[r.definiendum.decl.name]
         if decl.name not in self.def_vars:
@@ -145,15 +146,89 @@ def annotate(self, theory, voc, q_vars):
                 q_v[name] = Variable(name=name, sort=decl.out)
             self.def_vars[decl.name] = q_v
         new_rule = r.rename_args(self.def_vars[decl.name])
-        self.clarks.setdefault(decl, []).append(new_rule)
+        self.canonicals.setdefault(decl, []).append(new_rule)
 
     # join the bodies of rules
-    for decl, rules in self.clarks.items():
+    for decl, rules in self.canonicals.items():
+        new_rule = copy(rules[0])
         exprs = [rule.body for rule in rules]
-        rules[0].body = ADisjunction.make('∨', exprs)
-        self.clarks[decl] = rules[0]
+        new_rule.body = ADisjunction.make('∨', exprs)
+        self.clarks[decl] = new_rule
+    self.instantiables = self.get_instantiables()
     return self
 Definition.annotate = annotate
+
+def get_instantiables(self, for_explain=False):
+    """ compute Definition.instantiables, with level-mapping if definition is inductive
+
+    Uses implications instead of equivalence if `for_explain` is True
+
+    Example: `{ p() <- q(). p() <- r().}`
+    Result when not for_explain: `p() <=> q() | r()`
+    Result when for_explain    : `p() <= q(). p() <= r(). p() => (q() | r()).`
+
+    Args:
+        for_explain (Bool):
+            Use implications instead of equivalence, for rule-specific explanations
+    """
+    result = {}
+    for decl, rules in self.canonicals.items():
+        rule = rules[0]
+        if not rule.is_whole_domain:
+            self.check(rule.definiendum.symbol.decl not in self.level_symbols,
+                       f"Cannot have inductive definitions on infinite domain")
+        else:
+            if rule.out:
+                expr = AppliedSymbol.make(rule.definiendum.symbol,
+                                        rule.definiendum.sub_exprs[:-1])
+                expr.in_head = True
+                head = AComparison.make('=', [expr, rule.definiendum.sub_exprs[-1]])
+            else:
+                head = AppliedSymbol.make(rule.definiendum.symbol,
+                                        rule.definiendum.sub_exprs)
+                head.in_head = True
+
+            inductive = (not rule.out and DEF_SEMANTICS != Semantics.COMPLETION
+                and rule.definiendum.symbol.decl in rule.parent.level_symbols)
+
+            # determine reverse implications, if any
+            bodies, out = [], []
+            for r in rules:
+                if not inductive:
+                    bodies.append(r.body)
+                    if for_explain and 1 < len(rules):  # not simplified -> no need to make copies
+                        out.append(ARImplication.make('⇐', [head, r.body],
+                                                      r.annotations))
+                else:
+                    new = r.body.split_equivalences()
+                    bodies.append(new)
+                    if for_explain:
+                        new = new.add_level_mapping(rule.parent.level_symbols,
+                                             rule.definiendum, False, False)
+                        out.append(ARImplication.make('⇐', [head, new],
+                                                      r.annotations))
+
+            all_bodies = ADisjunction.make('∨', bodies)
+            if not inductive:
+                if out:  # already contains reverse implications
+                    out.append(ARImplication.make('⇒', [head, all_bodies],
+                                                  self.annotations))
+                else:
+                    out = [AEquivalence.make('⇔', [head, all_bodies],
+                                             self.annotations)]
+            else:
+                if not out:  # no reverse implication yet
+                    new = all_bodies.copy().add_level_mapping(rule.parent.level_symbols,
+                                             rule.definiendum, False, False)
+                    out = [ARImplication.make('⇐', [head.copy(), new],
+                                              self.annotations)]
+                all_bodies = all_bodies.add_level_mapping(rule.parent.level_symbols,
+                                        rule.definiendum, True, True)
+                out.append(AImplication.make('⇒', [head, all_bodies],
+                                             self.annotations))
+            result[decl] = out
+    return result
+Definition.get_instantiables = get_instantiables
 
 
 # Class Rule  #######################################################
@@ -174,10 +249,36 @@ def annotate(self, voc, q_vars):
     self.definiendum = self.definiendum.annotate(voc, q_v)
     self.body = self.body.annotate(voc, q_v)
 
-    self.is_whole_domain = all(s.name not in [INT, REAL, DATE]
+    self.is_whole_domain = all(s.name not in [INT, REAL, DATE]  # can't use s.range yet
                                for s in self.definiendum.decl.sorts)
     return self
 Rule.annotate = annotate
+
+def rename_args(self, new_vars):
+    """ for Clark's completion
+        input : '!v: f(args) <- body(args)'
+        output: '!nv: f(nv) <- nv=args & body(args)'
+    """
+    self.check(len(self.definiendum.sub_exprs) == len(new_vars), "Internal error")
+    vars = [var.name for q in self.quantees for vars in q.vars for var in vars]
+    for i in range(len(self.definiendum.sub_exprs)):
+        arg, nv = self.definiendum.sub_exprs[i], list(new_vars.values())[i]
+        if type(arg) == Variable \
+        and arg.name in vars and arg.name not in new_vars:
+            self.body = self.body.instantiate([arg], [nv])
+            self.out = (self.out.instantiate([arg], [nv]) if self.out else
+                        self.out)
+            for j in range(i, len(self.definiendum.sub_exprs)):
+                self.definiendum.sub_exprs[j] = \
+                    self.definiendum.sub_exprs[j].instantiate([arg], [nv])
+        else:
+            eq = AComparison.make('=', [nv, arg])
+            self.body = AConjunction.make('∧', [eq, self.body])
+
+    self.definiendum.sub_exprs = list(new_vars.values())
+    self.quantees = [Quantee.make(v, v.sort) for v in new_vars.values()]
+    return self
+Rule.rename_args = rename_args
 
 
 # Class Structure  #######################################################